Refactor constactMsgAction test
The test some time doesn`t past and other it do, becouse it doesn`t
found the error messages. That becouse we had been holdin elements
references before the await in the test, a some time the references
doesn`t work as expected. Now we all getting the elements inside the
userEvent and expect functions directly